+++ fabric-transfer-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/transfer/v1/storage/base/ResourceAmount.java
@@ -1,0 +1,5 @@
+package net.fabricmc.fabric.api.transfer.v1.storage.base;
+
+/** Immutable resource and amount pair. */
+public record ResourceAmount<T>(T resource, long amount) {
+}